refactor: extracted SendInput into separate file
